+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
     - `EasyCaptcha::Captcha`
     - `EasyCaptcha::ModelHelpers`
+    - `EasyCaptcha::Espeak` WAV file generation
 
 * [karlwilbur] - Add `mdl` gem dev dependency; run linters using `bundle exec`
 * [karlwilbur] - Add `CHANGELOG`